giggel/requirements.txt
confusable-homoglyphs==3.2.0
coverage==4.5.3
Django==2.2.2
django-crispy-forms==1.7.2
django-debug-toolbar==2.0
django-filter==2.2.0
django-registration==3.0.1
mysqlclient==1.4.4
Pillow==6.1.0
pytz==2019.1
selenium==3.141.0
sqlparse==0.3.0
urllib3==1.25.3
